requirements.dev.txt
flask

matplotlib
pandas
tensorflow-datasets

black
ipykernel